+++ pkgs/test_api/lib/src/backend/remote_exception.dart
@@ -12,7 +12,7 @@ import 'test_failure.dart';
 ///
 /// This could be an exception thrown in a different isolate, a different
 /// process, or on an entirely different computer.
-class RemoteException implements Exception {
+final class RemoteException implements Exception {
   /// The original exception's message, if it had one.
   ///
   /// If the original exception was a plain string, this will contain that
@@ -85,7 +85,7 @@ class RemoteException implements Exception {
 ///
 /// It's important to preserve [TestFailure]-ness, because tests have different
 /// results depending on whether an exception was a failure or an error.
-class _RemoteTestFailure extends RemoteException implements TestFailure {
+final class _RemoteTestFailure extends RemoteException implements TestFailure {
   _RemoteTestFailure(String? message, String type, String toString)
       : super._(message, type, toString);
 }