+++ pkgs/test_core/lib/src/runner/configuration/runtime_settings.dart
@@ -8,7 +8,7 @@ import 'package:yaml/yaml.dart';
 import '../plugin/customizable_platform.dart';
 
 /// User-defined settings for a built-in test runtime.
-class RuntimeSettings {
+final class RuntimeSettings {
   /// The identifier used to look up the runtime being overridden.
   final String identifier;