+++ pkgs/test/lib/src/runner/browser/default_settings.dart
@@ -13,7 +13,15 @@ final defaultSettings = UnmodifiableMapView({
       linuxExecutable: 'google-chrome',
       macOSExecutable:
           '/Applications/Google Chrome.app/Contents/MacOS/Google Chrome',
-      windowsExecutable: r'Google\Chrome\Application\chrome.exe'),
+      windowsExecutable: r'Google\Chrome\Application\chrome.exe',
+      environmentOverride: 'CHROME_EXECUTABLE'),
+  Runtime.edge: ExecutableSettings(
+    linuxExecutable: 'microsoft-edge-stable',
+    windowsExecutable: r'Microsoft\Edge\Application\msedge.exe',
+    macOSExecutable:
+        '/Applications/Microsoft Edge.app/Contents/MacOS/Microsoft Edge',
+    environmentOverride: 'MS_EDGE_EXECUTABLE',
+  ),
   Runtime.firefox: ExecutableSettings(
       linuxExecutable: 'firefox',
       macOSExecutable: '/Applications/Firefox.app/Contents/MacOS/firefox-bin',
@@ -21,5 +29,5 @@ final defaultSettings = UnmodifiableMapView({
   Runtime.internetExplorer:
       ExecutableSettings(windowsExecutable: r'Internet Explorer\iexplore.exe'),
   Runtime.safari: ExecutableSettings(
-      macOSExecutable: '/Applications/Safari.app/Contents/MacOS/Safari')
+      macOSExecutable: '/Applications/Safari.app/Contents/MacOS/Safari'),
 });