+++ pkgs/test/test/runner/browser/microsoft_edge_test.dart
@@ -1,10 +1,11 @@
-// Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-@TestOn('vm')
-@Tags(['firefox'])
 
-import 'package:test/src/runner/browser/firefox.dart';
+@TestOn('vm')
+@Tags(['edge'])
+
+import 'package:test/src/runner/browser/microsoft_edge.dart';
 import 'package:test/src/runner/executable_settings.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
@@ -16,7 +17,7 @@ import 'code_server.dart';
 void main() {
   setUpAll(precompileTestExecutable);
 
-  test('starts Firefox with the given URL', () async {
+  test('starts edge with the given URL', () async {
     var server = await CodeServer.start();
 
     server.handleJavaScript('''
@@ -27,29 +28,22 @@ webSocket.addEventListener("open", function() {
 ''');
     var webSocket = server.handleWebSocket();
 
-    var firefox = Firefox(server.url);
-    addTearDown(() => firefox.close());
+    var edge = MicrosoftEdge(server.url, configuration());
+    addTearDown(() => edge.close());
 
     expect(await (await webSocket).stream.first, equals('loaded!'));
-  });
-
-  test("a process can be killed synchronously after it's started", () async {
-    var server = await CodeServer.start();
-
-    var firefox = Firefox(server.url);
-    await firefox.close();
-  });
+  }, timeout: Timeout.factor(2));
 
   test('reports an error in onExit', () {
-    var firefox = Firefox('http://dart-lang.org',
+    var edge = MicrosoftEdge(Uri.parse('https://dart.dev'), configuration(),
         settings: ExecutableSettings(
             linuxExecutable: '_does_not_exist',
             macOSExecutable: '_does_not_exist',
             windowsExecutable: '_does_not_exist'));
     expect(
-        firefox.onExit,
+        edge.onExit,
         throwsA(isApplicationException(
-            startsWith('Failed to run Firefox: $noSuchFileMessage'))));
+            startsWith('Failed to run Edge: $noSuchFileMessage'))));
   });
 
   test('can run successful tests', () async {
@@ -61,7 +55,7 @@ void main() {
 }
 ''').create();
 
-    var test = await runTest(['-p', 'firefox', 'test.dart']);
+    var test = await runTest(['-p', 'edge', 'test.dart']);
     expect(test.stdout, emitsThrough(contains('+1: All tests passed!')));
     await test.shouldExit(0);
   });
@@ -75,26 +69,8 @@ void main() {
 }
 ''').create();
 
-    var test = await runTest(['-p', 'firefox', 'test.dart']);
+    var test = await runTest(['-p', 'edge', 'test.dart']);
     expect(test.stdout, emitsThrough(contains('-1: Some tests failed.')));
     await test.shouldExit(1);
-  });
-
-  test('not impacted by CHROME_EXECUTABLE var', () async {
-    await d.file('test.dart', '''
-import 'dart:html';
-
-import 'package:test/test.dart';
-
-void main() {
-  test("success", () {
-    assert(window.navigator.vendor != 'Google Inc.');
-  });
-}
-''').create();
-    var test = await runTest(['-p', 'firefox', 'test.dart'],
-        environment: {'CHROME_EXECUTABLE': '/some/bad/path'});
-    expect(test.stdout, emitsThrough(contains('+1: All tests passed!')));
-    await test.shouldExit(0);
   });
 }