+++ pkgs/test_core/lib/src/runner/hybrid_listener.dart
@@ -9,7 +9,6 @@ import 'package:async/async.dart';
 import 'package:stack_trace/stack_trace.dart';
 import 'package:stream_channel/isolate_channel.dart';
 import 'package:stream_channel/stream_channel.dart';
-// ignore: deprecated_member_use
 import 'package:test_api/backend.dart' show RemoteException;
 import 'package:test_api/src/utils.dart'; // ignore: implementation_imports