+++ pkgs/test_core/lib/src/runner/plugin/remote_platform_helpers.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:stream_channel/stream_channel.dart';
-// ignore: deprecated_member_use
 import 'package:test_api/backend.dart'
     show RemoteListener, StackTraceFormatter, StackTraceMapper;