+++ pkgs/checks/CHANGELOG.md
@@ -1,7 +1,15 @@
-## 0.2.2-dev
+## 0.2.3-wip
 
+-   Add class modifiers to restrict extension of implementation classes.
+
+## 0.2.2
+
+-   Return the first failure from `softCheck` and `softCheckAsync` as
+    documented, instead of the last failure when there are multiple failures.
 -   Add example `because` usage and mention the "reason" name in the migration
     guide.
+-   Add `ComparableChecks` with comparison expectations for subject types that
+    implement `Comparable`.
 
 ## 0.2.1